Move ipp cluster into conftest.py
